生成 Sql ,以及考虑到时间类型，速度不是很快

+++ src/main/java/trybin/fdg/entity/Columns.java
@@ -2,6 +2,8 @@ package trybin.fdg.entity;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * @author: TryBin
  * @date: 2021/10/28 18:15:18
@@ -10,9 +12,9 @@ import lombok.Data;
 @Data
 public class Columns {
     private String colname;
-    private Integer keyseq;
+    private Long keyseq;
     private String typename;
-    private Integer length;
-    private Integer scale;
+    private BigInteger length;
+    private Long scale;
     private String idEntity;
 }
